object_id: move HexPrefix and PrefixResolution from index module
