+++ lib/src/object_id.rs
@@ -79,3 +79,144 @@ macro_rules! impl_id_type {
 }
 
 pub(crate) use {id_type, impl_id_type};
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct HexPrefix {
+    // For odd-length prefix, lower 4 bits of the last byte is padded with 0
+    min_prefix_bytes: Vec<u8>,
+    has_odd_byte: bool,
+}
+
+impl HexPrefix {
+    pub fn new(prefix: &str) -> Option<HexPrefix> {
+        let has_odd_byte = prefix.len() & 1 != 0;
+        let min_prefix_bytes = if has_odd_byte {
+            hex::decode(prefix.to_owned() + "0").ok()?
+        } else {
+            hex::decode(prefix).ok()?
+        };
+        Some(HexPrefix {
+            min_prefix_bytes,
+            has_odd_byte,
+        })
+    }
+
+    pub fn from_bytes(bytes: &[u8]) -> Self {
+        HexPrefix {
+            min_prefix_bytes: bytes.to_owned(),
+            has_odd_byte: false,
+        }
+    }
+
+    pub fn hex(&self) -> String {
+        let mut hex_string = hex::encode(&self.min_prefix_bytes);
+        if self.has_odd_byte {
+            hex_string.pop().unwrap();
+        }
+        hex_string
+    }
+
+    /// Minimum bytes that would match this prefix. (e.g. "abc0" for "abc")
+    ///
+    /// Use this to partition a sorted slice, and test `matches(id)` from there.
+    pub fn min_prefix_bytes(&self) -> &[u8] {
+        &self.min_prefix_bytes
+    }
+
+    /// Returns the bytes representation if this prefix can be a full id.
+    pub fn as_full_bytes(&self) -> Option<&[u8]> {
+        (!self.has_odd_byte).then_some(&self.min_prefix_bytes)
+    }
+
+    fn split_odd_byte(&self) -> (Option<u8>, &[u8]) {
+        if self.has_odd_byte {
+            let (&odd, prefix) = self.min_prefix_bytes.split_last().unwrap();
+            (Some(odd), prefix)
+        } else {
+            (None, &self.min_prefix_bytes)
+        }
+    }
+
+    pub fn matches<Q: ObjectId>(&self, id: &Q) -> bool {
+        let id_bytes = id.as_bytes();
+        let (maybe_odd, prefix) = self.split_odd_byte();
+        if id_bytes.starts_with(prefix) {
+            if let Some(odd) = maybe_odd {
+                matches!(id_bytes.get(prefix.len()), Some(v) if v & 0xf0 == odd)
+            } else {
+                true
+            }
+        } else {
+            false
+        }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum PrefixResolution<T> {
+    NoMatch,
+    SingleMatch(T),
+    AmbiguousMatch,
+}
+
+impl<T> PrefixResolution<T> {
+    pub fn map<U>(self, f: impl FnOnce(T) -> U) -> PrefixResolution<U> {
+        match self {
+            PrefixResolution::NoMatch => PrefixResolution::NoMatch,
+            PrefixResolution::SingleMatch(x) => PrefixResolution::SingleMatch(f(x)),
+            PrefixResolution::AmbiguousMatch => PrefixResolution::AmbiguousMatch,
+        }
+    }
+}
+
+impl<T: Clone> PrefixResolution<T> {
+    pub fn plus(&self, other: &PrefixResolution<T>) -> PrefixResolution<T> {
+        match (self, other) {
+            (PrefixResolution::NoMatch, other) => other.clone(),
+            (local, PrefixResolution::NoMatch) => local.clone(),
+            (PrefixResolution::AmbiguousMatch, _) => PrefixResolution::AmbiguousMatch,
+            (_, PrefixResolution::AmbiguousMatch) => PrefixResolution::AmbiguousMatch,
+            (PrefixResolution::SingleMatch(_), PrefixResolution::SingleMatch(_)) => {
+                PrefixResolution::AmbiguousMatch
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::backend::CommitId;
+
+    #[test]
+    fn test_hex_prefix_prefixes() {
+        let prefix = HexPrefix::new("").unwrap();
+        assert_eq!(prefix.min_prefix_bytes(), b"");
+
+        let prefix = HexPrefix::new("1").unwrap();
+        assert_eq!(prefix.min_prefix_bytes(), b"\x10");
+
+        let prefix = HexPrefix::new("12").unwrap();
+        assert_eq!(prefix.min_prefix_bytes(), b"\x12");
+
+        let prefix = HexPrefix::new("123").unwrap();
+        assert_eq!(prefix.min_prefix_bytes(), b"\x12\x30");
+    }
+
+    #[test]
+    fn test_hex_prefix_matches() {
+        let id = CommitId::from_hex("1234");
+
+        assert!(HexPrefix::new("").unwrap().matches(&id));
+        assert!(HexPrefix::new("1").unwrap().matches(&id));
+        assert!(HexPrefix::new("12").unwrap().matches(&id));
+        assert!(HexPrefix::new("123").unwrap().matches(&id));
+        assert!(HexPrefix::new("1234").unwrap().matches(&id));
+        assert!(!HexPrefix::new("12345").unwrap().matches(&id));
+
+        assert!(!HexPrefix::new("a").unwrap().matches(&id));
+        assert!(!HexPrefix::new("1a").unwrap().matches(&id));
+        assert!(!HexPrefix::new("12a").unwrap().matches(&id));
+        assert!(!HexPrefix::new("123a").unwrap().matches(&id));
+    }
+}
