+++ lib/tests/test_id_prefix.rs
@@ -15,9 +15,8 @@
 use itertools::Itertools;
 use jj_lib::backend::{CommitId, MillisSinceEpoch, Signature, Timestamp};
 use jj_lib::id_prefix::IdPrefixContext;
-use jj_lib::index::HexPrefix;
-use jj_lib::index::PrefixResolution::{AmbiguousMatch, NoMatch, SingleMatch};
-use jj_lib::object_id::ObjectId;
+use jj_lib::object_id::PrefixResolution::{AmbiguousMatch, NoMatch, SingleMatch};
+use jj_lib::object_id::{HexPrefix, ObjectId};
 use jj_lib::repo::Repo;
 use jj_lib::revset::RevsetExpression;
 use testutils::{TestRepo, TestRepoBackend};
