+++ lib/tests/test_revset.rs
@@ -25,8 +25,7 @@ use jj_lib::backend::{ChangeId, CommitId, MillisSinceEpoch, Signature, Timestamp
 use jj_lib::commit::Commit;
 use jj_lib::git;
 use jj_lib::git_backend::GitBackend;
-use jj_lib::index::{HexPrefix, PrefixResolution};
-use jj_lib::object_id::ObjectId;
+use jj_lib::object_id::{HexPrefix, ObjectId, PrefixResolution};
 use jj_lib::op_store::{RefTarget, RemoteRef, RemoteRefState, WorkspaceId};
 use jj_lib::repo::Repo;
 use jj_lib::repo_path::RepoPath;
